+++ packages/smithy-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 4.11.1
+
+### Patch Changes
+
+- Updated dependencies [472bf01]
+  - @smithy/core@3.22.0
+  - @smithy/middleware-endpoint@4.4.12
+
+## 4.11.0
+
+### Minor Changes
+
+- 75145e5: add paginators/waiters to aggregate clients
+
 ## 4.10.12
 
 ### Patch Changes